+++ src/compiler/utilities/mod.rs
@@ -4,8 +4,10 @@ mod lines_1000_1500;
 pub use lines_1000_1500::*;
 mod lines_1500_2000;
 pub use lines_1500_2000::*;
-mod lines_2000_3000;
-pub use lines_2000_3000::*;
+mod lines_2000_2500;
+pub use lines_2000_2500::*;
+mod lines_2500_3000;
+pub use lines_2500_3000::*;
 mod lines_3000_4000;
 pub use lines_3000_4000::*;
 mod lines_4000_5000;
