+++ src/lib.rs
@@ -271,18 +271,20 @@ pub use compiler::utilities::{
     is_access_expression, is_ambient_module, is_any_import_or_re_export, is_assignment_expression,
     is_assignment_operator, is_bindable_static_access_expression,
     is_bindable_static_element_access_expression, is_block_or_catch_scoped, is_dynamic_name,
-    is_entity_name_expression, is_expression_with_type_arguments_in_class_extends_clause,
-    is_external_or_common_js_module, is_function_block, is_function_expression_or_arrow_function,
-    is_import_call, is_in_js_file, is_in_jsdoc, is_jsdoc_type_alias, is_json_source_file,
-    is_keyword, is_logical_or_coalescing_assignment_operator, is_object_literal_method,
-    is_prologue_directive, is_property_access_entity_name_expression, is_property_name_literal,
-    is_prototype_access, is_string_or_numeric_literal_like, is_super_property, is_this_identifier,
-    is_type_alias, is_type_node_kind, is_variable_like, is_write_only_access, modifier_to_flag,
+    is_effective_module_declaration, is_entity_name_expression,
+    is_expression_with_type_arguments_in_class_extends_clause, is_external_or_common_js_module,
+    is_function_block, is_function_expression_or_arrow_function, is_import_call, is_in_js_file,
+    is_in_jsdoc, is_jsdoc_type_alias, is_json_source_file, is_keyword,
+    is_logical_or_coalescing_assignment_operator, is_object_literal_method, is_prologue_directive,
+    is_property_access_entity_name_expression, is_property_name_literal, is_prototype_access,
+    is_string_or_numeric_literal_like, is_super_property, is_this_identifier, is_type_alias,
+    is_type_node_kind, is_variable_like, is_write_only_access, modifier_to_flag,
     modifiers_to_flags, node_is_missing, node_is_present, node_is_synthesized, object_allocator,
     parse_pseudo_big_int, position_is_synthesized, pseudo_big_int_to_string, set_parent,
     set_parent_recursive, set_text_range_pos, set_text_range_pos_end, set_text_range_pos_width,
-    set_value_declaration, using_single_line_string_writer, walk_up_parenthesized_expressions,
-    Associativity, FunctionFlags, GetLiteralTextFlags, OperatorPrecedence,
+    set_value_declaration, skip_parentheses, try_get_import_from_module_specifier,
+    using_single_line_string_writer, walk_up_parenthesized_expressions, Associativity,
+    FunctionFlags, GetLiteralTextFlags, OperatorPrecedence,
 };
 use compiler::utilities::{
     get_element_or_property_access_argument_expression_or_name,
